switch key to mongoId

+++ src/main/java/org/yajug/users/service/MemberServiceImpl.java
@@ -12,6 +12,7 @@ import org.slf4j.LoggerFactory;
 import org.yajug.users.domain.Member;
 import org.yajug.users.domain.Membership;
 import org.yajug.users.domain.Role;
+import org.yajug.users.domain.utils.KeyValidator;
 import org.yajug.users.persistence.dao.MemberMongoDao;
 import org.yajug.users.persistence.dao.MembershipMongoDao;
 
@@ -96,7 +97,7 @@ public class MemberServiceImpl implements MemberService {
 			
 			//expression validated against a pattern
 			if(!validateSearchExpression(expression)){
-				throw new DataException("invalid search pattern");
+				throw new ValidationException("Invalid search pattern");
 			}
 			members = memberMongoDao.search(expression);
 		}
@@ -111,7 +112,7 @@ public class MemberServiceImpl implements MemberService {
 			
 			//expression validated against a pattern
 			if(!validateSearchExpression(expression)){
-				throw new DataException("invalid search pattern");
+				throw new ValidationException("Invalid search pattern");
 			}
 			companies = memberMongoDao.getCompanies(expression);
 		}
@@ -123,23 +124,21 @@ public class MemberServiceImpl implements MemberService {
 	 */
 	@Override
 	public Collection<Membership> getMemberships(Member member) throws DataException {
-		Collection<Membership> memberships = new ArrayList<>();
-		if(member != null && member.getKey() > 0){
-			memberships = membershipMongoDao.getAllByMember(member.getKey());
+		if(member == null || !KeyValidator.validate(member.getKey())){
+			throw new ValidationException("Unable to retrieve a membership from an empty or invalid member");
 		}
-		return memberships;
+		return  membershipMongoDao.getAllByMember(member.getKey());
 	}
 	
 	/**
 	 * {@inheritDoc}
 	 */
 	@Override
-	public Membership getMembership(long id) throws DataException {
-		Membership membership = null;
-		if(id > 0){
-			membership = membershipMongoDao.getOne(id);
+	public Membership getMembership(String key) throws DataException {
+		if (!KeyValidator.validate(key)) {
+			throw new ValidationException("Unable to retrieve a membership from an empty or invalid key");
 		}
-		return membership;
+		return  membershipMongoDao.getOne(key);
 	}
 	
 	/**
@@ -149,7 +148,7 @@ public class MemberServiceImpl implements MemberService {
 	public boolean save(Member member) throws DataException{
 		
 		if(member == null){
-			throw new DataException("Cannot save a null member");
+			throw new ValidationException("Cannot save a null member");
 		}
 		
 		Collection<Member> members = new ArrayList<Member>();
@@ -165,19 +164,21 @@ public class MemberServiceImpl implements MemberService {
 	public boolean save(Collection<Member> members) throws DataException{
 		
 		if(members == null){
-			throw new DataException("Cannot save null members");
+			throw new ValidationException("Cannot save null members");
 		}
 		
 		int expected = members.size();
 		int saved = 0;
 		for(Member member : members){
-			if(memberMongoDao.isNew(member)){
-				if(memberMongoDao.insert(member)){
-					saved++;
-				}
-			} else {
-				if(memberMongoDao.update(member)){
-					saved++;
+			if(member != null){
+				if(memberMongoDao.isNew(member)){
+					if(memberMongoDao.insert(member)){
+						saved++;
+					}
+				} else if(KeyValidator.validate(member.getKey())) {
+					if(memberMongoDao.update(member)){
+						saved++;
+					}
 				}
 			}
 		}
@@ -188,32 +189,34 @@ public class MemberServiceImpl implements MemberService {
 	public boolean saveMemberships(Collection<Membership> memberships) throws DataException {
 		
 		if(memberships == null){
-			throw new DataException("Cannot save null memberships");
+			throw new ValidationException("Cannot save null memberships");
 		}
 		
 		//do the checks prior the save operations in a non transactionnal context
 		for(Membership membership : memberships){
 			if(membership == null){
-				throw new DataException("Can't save a null membership!");
+				throw new ValidationException("Can't save a null membership!");
 			}
 			if(membership.getYear() < 1990 || membership.getYear() > 2050){
-				throw new DataException("Wrong membership year : " + membership.getYear());
+				throw new ValidationException("Wrong membership year : " + membership.getYear());
 			}
 			if(membership.getType() == null){
-				throw new DataException("Membership type is required.");
+				throw new ValidationException("Membership type is required.");
 			}
 		}
 		
 		int expected = memberships.size();
 		int saved = 0;
 		for(Membership membership : memberships){
-			if(membershipMongoDao.isNew(membership)){
-				if(membershipMongoDao.insert(membership)){
-					saved++;
-				}
-			} else {
-				if(membershipMongoDao.update(membership)){
-					saved++;
+			if(membership != null){
+				if(membershipMongoDao.isNew(membership)){
+					if(membershipMongoDao.insert(membership)){
+						saved++;
+					}
+				} else if (KeyValidator.validate(membership.getKey())) {
+					if(membershipMongoDao.update(membership)){
+						saved++;
+					}
 				}
 			}
 		}
@@ -225,8 +228,8 @@ public class MemberServiceImpl implements MemberService {
 	 */
 	@Override
 	public boolean remove(Member member) throws DataException {
-		if(member == null){
-			throw new DataException("Cannot save a null member");
+		if(member == null || !KeyValidator.validate(member.getKey())){
+			throw new ValidationException("Cannot remove an empty or invalid member");
 		}
 		//remove also member's memberships
 		for(Membership membership : this.getMemberships(member)){
@@ -244,8 +247,8 @@ public class MemberServiceImpl implements MemberService {
 	 */
 	@Override
 	public boolean removeMembership(Membership membership) throws DataException {
-		if(membership == null){
-			throw new DataException("Cannot save a null member");
+		if(membership == null || !KeyValidator.validate(membership.getKey())){
+			throw new DataException("Cannot remove an empty or invalid membership");
 		}
 		return  membershipMongoDao.remove(membership);
 	}
